requirements.txt
pandas==1.5.3
numpy==1.21.6
scikit-learn==1.2.2
pyyaml==6.0
matplotlib==3.5.3
seaborn==0.11.2

notebook==6.5.4
jupyterlab==3.6.5
jupyter_server==1.21.0
jupyter_core==4.11.2
jupyter_client==7.4.8
traitlets==5.9.0
nbclassic==0.4.8

scipy>=1.9.0
statsmodels>=0.13.0

squarify>=0.4.3